+++ ingestor/app.py
@@ -4,12 +4,10 @@ from chalicelib import (
     s3_alerts,
     new_trains,
     bluebikes,
-    line_traversal,
+    daily_speeds,
     scheduled_speed,
-    dynamo,
-    update_agg_tables,
     constants,
-    populate_tables
+    agg_speed_tables
 )
 
 app = Chalice(app_name='ingestor')
@@ -62,8 +60,7 @@ def update_daily_speed_table(event):
     ''' Update yesterdays entry until 4/5 am (9 AM UTC)'''
     if today.hour < 9:
         today = today - timedelta(days=1)
-    line_traversal.update_daily_table(today)
-
+    daily_speeds.update_daily_table(today)
 
 
 # Runs every 5 minutes from either 4 AM -> 1:55AM or 5 AM -> 2:55 AM depending on DST
@@ -80,17 +77,16 @@ def update_scheduled_speed(event):
 # 7am UTC -> 2/3am EDT
 @app.schedule(Cron(0, 7, '*', '*', '?', '*'))
 def update_weekly_and_monthly_tables():
-    update_agg_tables.update_tables("weekly")
-    update_agg_tables.update_tables("monthly")
+    agg_speed_tables.update_tables("weekly")
+    agg_speed_tables.update_tables("monthly")
 
 
 @app.lambda_function()
 def populate_weekly_or_monthly_tables(params, context):
-    populate_tables.populate_table(params["line"], params["range"]) # monthly or weekly range
+    agg_speed_tables.populate_table(params["line"], params["range"]) # monthly or weekly range
 
 @app.lambda_function()
 def populate_daily(params, context):
     start_date = datetime.strptime("2016-01-15", constants.DATE_FORMAT_BACKEND)
     end_date = datetime.now()
-    line_traversal.populate_daily_table(start_date, end_date, params["line"])
-
+    daily_speeds.populate_daily_table(start_date, end_date, params["line"])
